Ensure that serial rx routines waiting until a terminator will complete
In the case of a serial port routine where we wait for a terminator, we
should eventually exit. This will handle the case where the terminator
is not received due to errors on the serial bus.

The hardware latency during exposure was measured, and found to be about
0.95 seconds. A large amount of timing slack was allowed for, just in
case this varies between cameras.