Improve the contrast of the JPEG images
To improve the contrast on the JPEG images, we substitute the CLAHE
algorithm for the percentile based method with a circular mask
as recommended by one of our astronomers.

This significantly improves the quality of the images taken during the
night. You can now clearly see stars in the images.