Refactor to reduce code duplication and add dark current support
We reduce the code duplication when saving files to any of the supported
formats. We also make the data structures which deal directly with the
camera simpler. An image processor is added which handles the conversion
from the raw ccd data provided by the camera into useful file formats.

The default module import surface is made much smaller, in an effort to
clarify how the library should be used.